[FEAT]: 심화기록 및 AI 태깅 api 연동
- today-task 캘린더 날짜 선택/스크럼 dot 상태 개선
- star-log 이미지 presigned 업로드 및 confirm 처리 추가
- AI 태깅 결과 조회 기반 성공 화면 렌더링
- 홈 요약 조회 및 리포트 모달 처리 추가
