+++ src/app/record/deep-log/page.tsx
@@ -166,7 +166,7 @@ const Page = () => {
       </section>
 
       {/* 이전 다음 버튼 영역 */}
-      <div className="flex shrink-0 gap-2 py-4">
+      <div className="pt-4, flex shrink-0 gap-2 pb-5">
         <Button
           size="lg"
           variant="gray"
