+++ src/containers/record/skill-tagging/SkillTaggingSuccess.tsx
@@ -7,99 +7,112 @@ import { useEffect, useState } from "react";
 import CTA from "@/components/common/CTA";
 import Modal from "@/components/common/Modal";
 import Tag from "@/components/common/Tag";
-import { RECORD_SKILL_TAGS } from "@/components/record/SkillTag";
-import FilledHeartGem from "@/components/record/stones/FilledHeartGem";
-import { mockReportGauge } from "@/data/report";
+import GlowingSkillStone, { type SkillStoneId } from "@/components/record/stones/GlowingSkillStone";
+import { PRIMARY_CATEGORY_MAP } from "@/constants/competency";
+import {
+  type AiTaggingResultResponse,
+  type Competency,
+  getHomeSummary,
+  type ReportModalType,
+} from "@/lib/apis/record/record";
 
-interface SkillTaggingTask {
-  id: number;
-  title: string;
-  projectId: number;
-  projectTag: string;
-  projectTitle: string;
-  skillId: number;
+const CATEGORY_STONE_ID: Record<Competency, SkillStoneId> = {
+  DISCOVERY_ANALYSIS: 1,
+  PLANNING_EXECUTION: 2,
+  COLLABORATION: 3,
+  PROBLEM_SOLVING: 4,
+  REFLECTION_GROWTH: 5,
+};
+
+function getPrimaryCategoryLabels(results: AiTaggingResultResponse[]) {
+  return Array.from(
+    new Set(
+      results.flatMap(result =>
+        result.primaryCategory
+          ? [PRIMARY_CATEGORY_MAP[result.primaryCategory]?.label ?? result.primaryCategory]
+          : [],
+      ),
+    ),
+  );
 }
 
-function getSkillIds(tasks: SkillTaggingTask[]) {
-  return Array.from(new Set(tasks.map(task => task.skillId)));
+function getDetailTagLabels(results: AiTaggingResultResponse[]) {
+  return Array.from(new Set(results.flatMap(result => result.detailTags ?? [])));
 }
 
-function getExperienceTagLabels(tasks: SkillTaggingTask[]) {
-  return Array.from(new Set(tasks.flatMap(task => [task.projectTag, task.title])));
+function getStoneIds(results: AiTaggingResultResponse[]) {
+  return Array.from(
+    new Set(
+      results.flatMap(result =>
+        result.primaryCategory ? [CATEGORY_STONE_ID[result.primaryCategory]] : [],
+      ),
+    ),
+  );
 }
 
-function getHeartItems(count: number) {
-  return Array.from({ length: Math.min(Math.max(count, 1), 20) }, (_, index) => ({
-    id: index,
-  }));
+function StonePreview({ stoneIds }: { stoneIds: SkillStoneId[] }) {
+  return (
+    <div className="flex max-w-48 flex-wrap justify-center gap-4.5">
+      {stoneIds.map(stoneId => (
+        <GlowingSkillStone key={stoneId} skillId={stoneId} animate className="size-17" />
+      ))}
+    </div>
+  );
 }
 
-function getReportMilestone(currentCount: number) {
-  if (currentCount < 10) return null;
-  if (currentCount % 10 !== 0) return null;
-
-  return currentCount / 10;
-}
-
-function FloatingHeart() {
-  return <FilledHeartGem ariaLabel="역량 태깅 성공" className="size-17" />;
-}
-
-function SkillTaggingSuccess({
-  tasks,
-  shouldCheckMiniReport,
-}: {
-  tasks: SkillTaggingTask[];
-  shouldCheckMiniReport: boolean;
-}) {
+function SkillTaggingSuccess({ results }: { results: AiTaggingResultResponse[] }) {
   const router = useRouter();
-  const [isReportModalOpen, setIsReportModalOpen] = useState(false);
-  const skillIds = getSkillIds(tasks);
-  const heartItems = getHeartItems(skillIds.length);
-  const reportMilestone = getReportMilestone(mockReportGauge.currentCount);
-  const isReportReady = shouldCheckMiniReport && reportMilestone !== null;
+  const primaryCategoryLabels = getPrimaryCategoryLabels(results);
+  const detailTagLabels = getDetailTagLabels(results);
+  const stoneIds = getStoneIds(results);
+  const [reportModalType, setReportModalType] = useState<ReportModalType | null>(null);
   const reportModalTitle =
-    reportMilestone === 1 ? "커리어 미니 리포트를 발행해보세요" : "커리어 리포트를 발행해보세요";
+    reportModalType === "MINI"
+      ? "커리어 미니 리포트를 발행해보세요"
+      : "커리어 리포트를 발행해보세요";
 
   useEffect(() => {
-    if (!isReportReady) return;
+    let ignore = false;
 
-    const reportModalTimer = window.setTimeout(() => {
-      setIsReportModalOpen(true);
-    }, 1500);
+    const loadHomeSummary = async () => {
+      try {
+        const summary = await getHomeSummary();
+
+        if (summary?.isFirstStar !== undefined) {
+          window.sessionStorage.setItem("isFirstStar", String(summary.isFirstStar));
+        }
+
+        if (!ignore && summary?.reportModal?.show && summary.reportModal.type) {
+          setReportModalType(summary.reportModal.type);
+        }
+      } catch {}
+    };
+
+    void loadHomeSummary();
 
     return () => {
-      window.clearTimeout(reportModalTimer);
+      ignore = true;
     };
-  }, [isReportReady]);
+  }, []);
 
   return (
     <section className="relative -mx-5 flex min-h-0 flex-1 flex-col overflow-hidden px-5">
-      {/* TODO: 추후 gif 파일로 이미지 변경 예정 */}
       <div className="relative flex min-h-0 flex-1 flex-col items-center justify-center">
-        <div className="flex max-w-48 flex-wrap justify-center gap-4.5">
-          {heartItems.map(heart => (
-            <FloatingHeart key={heart.id} />
-          ))}
-        </div>
+        <StonePreview stoneIds={stoneIds} />
 
         <div className="relative z-10 flex flex-col items-center">
           <div className="mt-3.75 flex flex-col items-center px-8">
             <p className="head-4 text-white">오늘의 경험이</p>
             <h1 className="head-4 text-sea-blue-500 text-center">
-              [
-              {skillIds
-                .map(skillId => RECORD_SKILL_TAGS.find(skill => skill.id === skillId)?.label)
-                .join("] [")}
-              ]
+              [{primaryCategoryLabels.join("] [")}]
             </h1>
             <p className="head-4 text-white">으로 기록됐어요</p>
           </div>
 
           <div className="scrollbar-hide mt-3 flex flex-wrap justify-center gap-1 overflow-y-auto px-3">
-            {getExperienceTagLabels(tasks).map(tagLabel => (
+            {detailTagLabels.map(tagLabel => (
               <Tag key={tagLabel} variant="gray" className="bg-gray-850">
-                #{tagLabel}
+                {tagLabel}
               </Tag>
             ))}
           </div>
@@ -111,22 +124,22 @@ function SkillTaggingSuccess({
         </Link>
       </div>
 
-      {/* 심화기록 10개 단위로 달성 시 팝업 모달 */}
       <Modal
-        isOpen={isReportModalOpen}
+        isOpen={reportModalType !== null}
         type="double"
         title={reportModalTitle}
         contents={
           <>
             지금까지 쌓인 기록으로 만들어진
             <br />
-            커리어 미니 리포트를 확인해보세요
+            커리어 {reportModalType === "MINI" ? "미니 " : ""}리포트를 확인해보세요
           </>
         }
         btnLLabel="다음에 보기"
         btnRLabel="리포트 만들기"
-        onBtnLClick={() => setIsReportModalOpen(false)}
+        onBtnLClick={() => setReportModalType(null)}
         onBtnRClick={() => router.push("/report")}
+        onClose={() => setReportModalType(null)}
         contentClassName="px-7 py-5"
         btnLClassName="px-5"
         btnRClassName="px-5"
