+++ src/lib/apis/record/record.ts
@@ -1,6 +1,6 @@
 import { api } from "@/api/client";
 
-type Competency =
+export type Competency =
   | "DISCOVERY_ANALYSIS"
   | "PLANNING_EXECUTION"
   | "COLLABORATION"
@@ -14,10 +14,20 @@ interface AiTaggingStatusResponse {
   retryCount?: number;
 }
 
-interface AiTaggingResultResponse {
+export interface AiTaggingResultResponse {
   status?: AiTaggingStatus;
   primaryCategory?: Competency;
   detailTags?: string[];
+}
+
+export type ReportModalType = "MINI" | "FULL";
+
+export interface HomeSummaryResponse {
+  isFirstStar?: boolean;
+  reportModal?: {
+    show?: boolean;
+    type?: ReportModalType | null;
+  };
 }
 
 // AI 태깅 트리거
@@ -31,3 +41,6 @@ export const getAiTaggingStatus = (starRecordId: number) =>
 // AI 태깅 결과 조회
 export const getAiTaggingResult = (starRecordId: number) =>
   api.get<AiTaggingResultResponse>(`/api/star-records/${starRecordId}/ai-tagging/result`);
+
+// 홈 복귀 요약 정보 조회
+export const getHomeSummary = () => api.get<HomeSummaryResponse>("/api/star-records/home-summary");
